Added SpriteSheets for jetpack and teleporting enemy
Each frame follows the standard 100 by 100px spacing.

Notes on the jetpack:
When drawing on character, offset jetpack's top left frame coordinates
to (-15,23) from the top left of the character's frame. The exception is
when character is kneeling, in which case, offset the jetpacks (-15,41)
from top left of character's frame. Also, if the jet packs are to appear
on the course before the character collects them, the bottom of the
jetpack is at the coordinates (50,50) within its 100x100 frame. This way
it will appear as if it's resting on the ground.